refactor(imageView): replace loop index type with size_t and extract image view creation to a separate method

+++ src/plugin/vk-wrapper/src/framebuffer/Framebuffer.cpp
@@ -6,7 +6,7 @@ void Framebuffer::Create(const VkDevice &device, const CreateInfo &info)
 {
     _swapChainFramebuffers.resize(info.swapChainImageViews.size());
 
-    for (uint32_t i = 0; i < info.swapChainImageViews.size(); ++i)
+    for (size_t i = 0; i < info.swapChainImageViews.size(); ++i)
     {
         std::array<VkImageView, 1> attachments = {info.swapChainImageViews[i]};
